Add files via upload
# Using satsuma2 to align congeneric species' assemblies at the scaffold level to chromosomal assemblies of the next most closely related species to improve the use of MSMC2 across taxa. 
Step 1: 01_rename_chrom_satsuma.sh renames chromosomes consistently across taxa. It will take the file "chrom_list.txt" which is a list of assemblies to be renamed. You should have a list of chromosomes for each assembly and what you want them to be renamed as. An example for the latter is provided as "chrom_example.txt". 

Step 2: 02_run_satsuma.sh. This step takes a list of target assemblies and a list of query assemblies that will align to the target assemblies. It will run satsuma2 and output a new assembly titled likeso "${query}_pseudo.fasta".